+++ src/backends/libtorch_c/torch_extensions/torch_redis.cpp
@@ -1,6 +1,6 @@
 #include <string>
 #include "torch_redis.h"
-#include "../../redismodule.h"
+#include "../../../redismodule.h"
 
 torch::IValue IValueFromRedisReply(RedisModuleCallReply *reply){
